Include syringe data file for BD Ultra-Fine U-100
Introduce data file for BD Ultra-Fine U-100 (EAN: 0382903261031) syringes which are commonly used for administering insulin in pets in Australia. Values were obtained by a combined average measurement of 10 needles selected from a batch of 100 for having accurately printed line markings, calibration tools were then printed to check with this test batch of needles, after which incremental changes were made to the head gap and head offset values to dial in the accuracy.